+++ package.json
@@ -18,12 +18,16 @@
     "hook",
     "react-hook"
   ],
+  "bin": {
+    "css-components": "lib/cli/index.js"
+  },
   "scripts": {
     "build": "tsc",
     "build:types": "tsc --emitDeclarationOnly",
     "prepublishOnly": "npm run build",
     "test": "jest --verbose",
     "test:watch": "jest --verbose --watch",
+    "test:cli": "npm run build && node lib/cli/index.js --css \"test/**/*.css\"",
     "coverage": "jest --coverage",
     "lint": "NODE_ENV=test npm-run-all --parallel lint:*",
     "lint:js": "eslint \"src/**/*.{js,jsx,ts,tsx}\"",
@@ -44,6 +48,7 @@
     "@babel/preset-typescript": "^7.18.6",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
+    "@types/glob": "^8.0.0",
     "@types/jest": "^29.2.2",
     "@types/react": "^18.0.25",
     "@typescript-eslint/eslint-plugin": "^5.42.1",
@@ -66,6 +71,8 @@
     "typescript": "^4.8.4"
   },
   "dependencies": {
+    "glob": "^8.0.3",
+    "glob-promise": "^5.0.0",
     "yargs": "^17.6.2"
   }
 }
